rover_transport: lint the serial, UDP and io_context packages; explain their QoS
rover_serial_driver, rover_udp_driver and rover_io_context had no ament_lint_auto block,
unlike rover_modbus_driver. They get the same block (uncrustify excluded for the 4-space
house style) and the same CPPLINT.cfg. What the linters then found is fixed:
- CMake continuation lines use 2-space indents;
- <url> moves before <author>, as package_format3 requires;
- two static_cast<long> become std::ptrdiff_t;
- one C header moves ahead of the C++ ones.

The bridge nodes' QoS choices now carry the same kind of rationale as the rest of the
codebase: reliable depth-100 reads (framed bytes / whole datagrams must not be dropped), and
best-effort write subscriptions. The best-effort part is on purpose: it matches any publisher,
and the only writer (rover_led's LED frames) streams data where the next packet supersedes a
dropped one.

+++ rover_transport/rover_udp_driver/test/integration/test_asio_udp_socket.cpp
@@ -22,14 +22,15 @@
 
 #include <gtest/gtest.h>
 
+#include <unistd.h>
+
 #include <chrono>
 #include <condition_variable>
+#include <cstddef>
 #include <cstdint>
 #include <mutex>
 #include <string>
 #include <vector>
-
-#include <unistd.h>
 
 #include "rover_udp_driver/infrastructure/asio_udp_socket.hpp"
 
@@ -100,7 +101,7 @@ TEST(AsioUdpSocketTest, AsyncSendReachesAnAsyncReceiver)
         [&](const std::vector<uint8_t> & buffer, std::size_t length)
         {
             std::lock_guard<std::mutex> lock{mutex};
-            received.assign(buffer.begin(), buffer.begin() + static_cast<long>(length));
+            received.assign(buffer.begin(), buffer.begin() + static_cast<std::ptrdiff_t>(length));
             got_it = true;
             cv.notify_one();
         });
